Passage en double des posX/PosY rides et riderequests

+++ smaaahh-web/Models/RideRequest.cs
@@ -14,10 +14,10 @@ namespace smaaahh_web
         
         public Rider Rider { get; set; }
 
-        public decimal PosXStart { get; set; }
-        public decimal PosYStart { get; set; }
-        public decimal PosXEnd { get; set; }
-        public decimal PosYEnd { get; set; }
+        public double PosXStart { get; set; }
+        public double PosYStart { get; set; }
+        public double PosXEnd { get; set; }
+        public double PosYEnd { get; set; }
 
         public int PlaceNumber { get; set; }
         public DateTime DateCreation { get; set; }
